refactor: move node fe views to views folder
