+++ src/node/views/composables/usePanelVisible.ts
