+++ src/node/views/utils/event.ts
